+++ Documentation/Developer/_Api/_MyController2.php
@@ -8,11 +8,16 @@ use Brotkrueml\Schema\Type\TypeFactory;
 
 final class MyController
 {
+    public function __construct(
+        private readonly TypeFactory $typeFactory,
+    ) {
+    }
+
     public function doSomething(): void
     {
         // ...
 
-        $person = TypeFactory::createType('Person');
+        $person = $this->typeFactory->create('Person');
         $person
             ->setId('https://example.org/#person-42')
             ->setProperty('givenName', 'John')