+++ Documentation/Developer/_Api/_MyController3.php
@@ -8,11 +8,16 @@ use Brotkrueml\Schema\Type\TypeFactory;
 
 final class MyController
 {
+    public function __construct(
+        private readonly TypeFactory $typeFactory,
+    ) {
+    }
+
     public function doSomething(): void
     {
         // ...
 
-        $event = TypeFactory::createType('Event')
+        $event = $this->typeFactory->create('Event')
             ->setProperty('name', 'Fancy Event')
             ->setProperty('image', 'https:/example.org/event.png')
             ->setProperty('url', 'https://example.org/')