+++ Documentation/Developer/_Index/_MyController1.php
@@ -8,11 +8,16 @@ use Brotkrueml\Schema\Type\TypeFactory;
 
 final class MyController
 {
+    public function __construct(
+        private readonly TypeFactory $typeFactory,
+    ) {
+    }
+
     public function doSomething(): void
     {
         // ...
 
-        $thing = TypeFactory::createType('Thing');
+        $thing = $this->typeFactory->create('Thing');
         $thing->setProperty('name', 'A thing');
 
         // ...