+++ Documentation/Developer/_MainEntity/_MyController.php
@@ -11,6 +11,7 @@ final class MyController
 {
     public function __construct(
         private readonly SchemaManager $schemaManager,
+        private readonly TypeFactory $typeFactory,
     ) {
     }
 
@@ -18,11 +19,11 @@ final class MyController
     {
         // ...
 
-        $aggregateRating = TypeFactory::createType('AggregateRating')
+        $aggregateRating = $this->typeFactory->create('AggregateRating')
             ->setProperty('ratingValue', '4')
             ->setProperty('reviewCount', '126');
 
-        $product = TypeFactory::createType('Product')
+        $product = $this->typeFactory->create('Product')
             ->setProperties([
                 'name' => 'Some fancy product',
                 'color' => 'blue',